// src/app/website/shared/models/profile.model.ts
export class Profile {
    // _id?: string;
    fname?: string;
    lname?: string;
    phoneno?: number;
    dateofbirth?: string;
    placeofbirth?: string;
    timeofbirth?: string;
    raasi?: string;
    nakshatram?: string;
    houseno?: string;
    street?: string;
    city?: string;
    pin?: number;
    state?: string;
    userid?: string;
  }